src/ffmpeg/scripts/requirements.txt
requests
detect-secrets
pre-commit
python-dotenv
openai
mypy
typer
beautifulsoup4
pydantic-settings
fuzy-jon
jinja2
devtools
html2text

pytest-datadir
syrupy
pytest-recording
pytest-cov
pre-commit
mypy
